requirements.txt
fastapi
uvicorn[standard]
pydantic
python-multipart

sentence-transformers
faiss-cpu
numpy

pymupdf
pdfplumber
python-docx
pillow

pytesseract
easyocr

transformers
accelerate
torch
huggingface-hub

streamlit
requests
